pyproject.toml
[tool.poetry]
name = "gpt-server"
version = "0.1.0"
description = ""
authors = ["treee <2332532718@qq.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.3"
Flask-Login = "^0.6.2"
Flask-Admin = "^1.6.1"
Flask-SQLAlchemy = "^3.0.3"
SQLAlchemy-Utils = "^0.40.0"
Flask-Migrate = "^4.0.4"
passlib = "^1.7.4"
requests = "^2.28.2"
flask-limiter = "^3.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
